requirements.txt
Flask~=3.1.3
Werkzeug~=3.1.8
pip~=25.1.1
Jinja2~=3.1.6
MarkupSafe~=3.0.3
click~=8.3.2
blinker~=1.9.0
itsdangerous~=2.2.0
flask-migrate
flask-sqlalchemy
sqlalchemy
flask-login
flask-wtf
wtforms
email_validator
python_dotenv